Refactor get headers and cookies
